TaskStream-0 fix: fix remarks
