+++ ServiceTests/ClientsServicesTests.cs
@@ -3,8 +3,6 @@ using Services.Exceptions;
 using Services;
 using Services.Storages;
 using Xunit;
-using ExportTool;
-using ModelsDb;
 
 namespace ServiceTests
 {
@@ -217,60 +215,6 @@ namespace ServiceTests
             {
                 Assert.True(false);
             }
-        }
-
-        [Fact]
-        public void ExportClientPositiveTest()
-        {
-            //Arrange
-            string pathToDirectory = Path.Combine("C:", "Users", "user", "source", "repos",
-                ".net-course-2022-Stadnic", "Tools", "TestFiles");
-            string fileName = "clients.csv";
-            ClientExporter clientExporter = new ClientExporter(pathToDirectory, fileName);
-
-            var clientStorage = new ClientStorage();
-            var clientService = new ClientService(clientStorage);
-            var clientsDB = clientStorage.Data.Clients.ToList();
-            List<Client> clients = new List<Client>();
-            for (int i = 0; i < clientsDB.Count; i++)
-            {
-                Guid id = clientsDB[i].Id;
-                var client = clientService.Get(id);
-                clients.Add(client);                
-            }
-
-            //Act            
-            clientExporter.WriteClientToCsv(clients);
-            var clientsFromFile = clientExporter.ReadClientFromCsv();
-
-            //Assert
-            Assert.Equal(clients[0].Name, clientsFromFile[0].Name);            
-        }
-
-        [Fact]
-        public void AddClientFromCSVToDBPositiveTest()
-        {
-            //Arrange
-            string pathToDirectory = Path.Combine("C:", "Users", "user", "source", "repos",
-                ".net-course-2022-Stadnic", "Tools", "TestFiles");
-            string fileName = "clients.csv";
-            ClientExporter clientExporter = new ClientExporter(pathToDirectory, fileName);
-
-            var clientStorage = new ClientStorage();
-            var clientService = new ClientService(clientStorage);       
-            var clientsFromFile = clientExporter.ReadClientFromCsv();
-            int i = 0;
-
-            //Act
-            foreach (var client in clientsFromFile)
-            {
-                clientService.Add(client);
-                i++;
-            }
-            clientStorage.Data.SaveChanges();
-
-            //Assert
-            Assert.True(clientsFromFile.Count == i);
-        }
+        }        
     }
 }
